pkg/uprobetracer: move common binary helper functions to separate file
